JWT 중간 세이브

+++ src/main/java/com/example/demo/repository/MemberRepository.java
@@ -16,7 +16,7 @@ public interface MemberRepository {
 
     public void doLogin(int id);
 
-    public int getLastInsertId();
+    public Long getLastInsertId();
 
     public Member getMemberById(Long id);
